Put side-effects into a separate module
As discussed in https://github.com/w3c/i-slide/pull/41#discussion_r758589679 and following convention described in https://open-wc.org/guides/developing-components/publishing/#do-export-element-classes
