+++ test/resources/demo.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset=utf-8>
     <title>Demo of the &lt;i-slide&gt; Web component</title>
-    <script src="i-slide.js?register&selector=a.islide" type=module></script>
+    <script src="i-slide.js?selector=a.islide" type=module></script>
   </head>
   <body>
     <h1>Demo of the &lt;i-slide&gt; Web component</h1>
